migrate routes to ts
