+++ server/global.interface.ts
@@ -1,52 +1,15 @@
+import { IUser } from './models/User'
+
 declare global {
     namespace Express {
         interface Request {
-            user?: UserInterface
+            user?: IUser
         }
     }
 }
 
-interface Document<T> {
-    _doc: T
+export interface UserToReturn extends Omit<IUser, 'password'> {
+    password?: string
 }
 
-export interface UserInterface extends Document<UserInterface> {
-    role: string
-    username: string
-    email: string
-    password: string
-    image: string
-    color: string
-    level: number
-    xp: number
-    stats: UserStats
-    activeBattle: boolean
-    coin: number
-    inventory: Array<Item>
-    onboardingStage: number
-    createdAt: Date
-    updatedAt: Date
-}
-
-interface UserStats {
-    battles: number
-    wins: number
-    losses: number
-    draws: number
-}
-
-interface Item {
-    contents: ItemContents
-    image: string
-    info: string
-    level: number
-    name: string
-    price: number
-    type: string
-    _id: string
-}
-
-interface ItemContents {
-    count: number
-    odds: any
-}
+export type AuthenticatedUser = IUser & { password?: never }
