+++ server/package.json
@@ -23,6 +23,7 @@
         "validator": "^13.9.0"
     },
     "devDependencies": {
+        "@types/bcryptjs": "^2.4.2",
         "@types/cookie-parser": "^1.4.3",
         "@types/express": "^4.17.17",
         "@types/jsonwebtoken": "^9.0.2",
